refactor NVS and read WiFi parameters from NVS
The NVS API has been refactored and the Singleton implementation removed
for code simplicity (i don't say it isn't needed but when needed the
singleton can be made on another class), `SingleNVS` was replaced by
`Namespace`.

The Namespace class is a on object that wraps the NVS C namespace API,
it provides safe functions and automatically closes the namespace when
the object is dropped.