components: HTTP use esp_http_server component
Hand written code has been replaced with the `esp_http_server` component
from the ESP-IDF. Also, for parsing JSON we'll be using cJSON (provided
by the ESP-IDF too). The code is rough for now, some things are missing
and the REST API can be redesigned.

I would bet on the following:

- `/wifi/ap`, POST, Change AP settings.
- `/wifi/sta`, POST, Change STA settings.
- `/system/info`, GET, Get system information.

Co-authored-By: gustavosinbandera1 <gustavosinbandera1@hotmail.com>